A whole bunch of cleanup updates, and modification of renderhtmlreprot and rendercsvreport to match rendergridreport that much more closely.  This helps for the ability to print the pages nicely, as well as iterate across sub-sets in tables, both for printing and legibility.  It might work at some point, to distil down the renderfooreport functions to just one renderreport function, with three return values, those being html, csv, and grid.  Of course, we might be able to massage grid to match html at some point, as well, giving us just two options.
Also the correct_footer function was written, perhaps the correct_header function can be written as well, replacing all the variety of foo_footer and foo_header, since we have the switches.  These might also be able to be boiled down into just two functions instead of the nine or ten current ones.